Be more cautious about deleting a mailbox
First check to see if the mailbox is an alias recipient. If we
do the delete, we orphan the alias with a dead recipient. Return
an error instead forcing the user/app to back up and clean up.
Deleting the mailbox itself is the last thing you do.

Also, replace deleteAddress by and explicit delete of the address
row. deleteAddress quietly deletes the referenced domain row which
is fine for aliases but not so for mailboxes. We explicitly expect
the domain to already exist as part of the dovecot storage setup
so return the favor on individual mbox accounts. Let the admin
clean up the domain in both places.

Signed-off-by: Jim Lieb <lieb@sea-troll.net>