Fix PreRun calls
It seems that cobra may, someday, chase back and call
all the parent's preruns. But not yet.

Export and import have a preruns to do I/O redirection. This
is all well and good but the root prerun to open the DB does
not run by default. With this hook, now they do...

Signed-off-by: Jim Lieb <lieb@sea-troll.net>